scripts/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.9.0.0
pyyaml>=6.0.1